requirements.txt
bson==0.5.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
marshmallow==3.0.0b4
pkg-resources==0.0.0
pycrypto==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
speakin-voice-sdk==1.0.1
urllib3==1.22
